backend/requirements.txt
Flask
Flask-Cors
SQLAlchemy
psycopg2-binary  # or mysqlclient for MySQL
python-dotenv
selenium
